travelEye/requirements.txt
appdirs==1.4.3
Django==1.9
django-cors-headers==2.0.2
django-filter==1.0.2
django-leaflet==0.21.0
djangorestframework==3.5.4
djangorestframework-gis==0.11
Markdown==2.6.8
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
six==1.10.0
